feat: add floating CurrencyRatesWidget to MainLayout and remove legacy sidebar currency styles
