+++ src/components/general/CurrencyRatesWidget.jsx
@@ -1,0 +1,80 @@
+import React, { useState, useEffect, useMemo } from "react";
+import "../../styles/components/CurrencyRatesWidget.scss";
+import { fetchConversionRates } from "../../api/conversion/conversion.js";
+import {
+  buildCurrencyRateRows,
+  getCurrencyDisplayLabel,
+} from "../../api/utils/getCurrencyCode.js";
+
+const CURRENCY_META = {
+  USD: { flag: "🇺🇸", label: "Доллар" },
+  EUR: { flag: "🇪🇺", label: "Евро" },
+  RUB: { flag: "🇷🇺", label: "Рубль" },
+};
+
+const CurrencyRatesWidget = () => {
+  const [rates, setRates] = useState([]);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState(false);
+
+  useEffect(() => {
+    let mounted = true;
+    const load = async () => {
+      try {
+        const data = await fetchConversionRates(new Date());
+        if (mounted) {
+          setRates(data);
+          setLoading(false);
+        }
+      } catch {
+        if (mounted) {
+          setError(true);
+          setLoading(false);
+        }
+      }
+    };
+    load();
+    const interval = setInterval(load, 600000); // Update every 10 minutes
+
+    return () => {
+      mounted = false;
+      clearInterval(interval);
+    };
+  }, []);
+
+  const grouped = useMemo(() => {
+    const rows = buildCurrencyRateRows(rates).map((row) => ({
+      ...row,
+      meta: CURRENCY_META[row.currency],
+    }));
+    // Filter out currencies without meta info if needed, or keep all
+    return rows.filter(r => r.meta);
+  }, [rates]);
+
+  if (loading || error) return null; // Silently hide or show minimal state if preferred
+
+  return (
+    <div className="currency-rates-overlay">
+      {grouped.map((row) => (
+        <div key={row.currency} className="currency-card">
+          <div className="currency-card-left">
+            <div className="currency-flag-circle">
+              {row.meta.flag}
+            </div>
+            <div className="currency-code">{row.currency}</div>
+          </div>
+          <div className="currency-card-rates">
+            <span className="rate-value buy">
+              {row.buy != null ? row.buy.toFixed(4) : "—"}
+            </span>
+            <span className="rate-value sell">
+              {row.sell != null ? row.sell.toFixed(4) : "—"}
+            </span>
+          </div>
+        </div>
+      ))}
+    </div>
+  );
+};
+
+export default CurrencyRatesWidget;
