+++ src/styles/components/CurrencyRatesWidget.scss
@@ -1,0 +1,115 @@
+.currency-rates-overlay {
+  position: fixed;
+  bottom: 24px;
+  left: 24px;
+  display: flex;
+  gap: 12px;
+  z-index: 9999;
+  pointer-events: auto;
+  transition: all 0.3s ease;
+
+  // Responsive adjustment for mobile
+  @media (max-width: 768px) {
+    bottom: 16px;
+    left: 16px;
+    gap: 8px;
+    scale: 0.9;
+    transform-origin: bottom left;
+  }
+}
+
+.currency-card {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  padding: 12px 20px;
+  background: rgba(255, 255, 255, 0.7);
+  backdrop-filter: blur(12px);
+  -webkit-backdrop-filter: blur(12px);
+  border-radius: 20px;
+  border: 1px solid rgba(255, 255, 255, 0.4);
+  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.08);
+  transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+  min-width: 160px;
+
+  &:hover {
+    transform: translateY(-4px);
+    background: rgba(255, 255, 255, 0.85);
+    box-shadow: 0 12px 40px rgba(0, 0, 0, 0.12);
+  }
+
+  .currency-card-left {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+
+    .currency-flag-circle {
+      width: 32px;
+      height: 32px;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      background: white;
+      border-radius: 50%;
+      font-size: 20px;
+      box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
+    }
+
+    .currency-code {
+      font-size: 1.1rem;
+      font-weight: 700;
+      color: #1a1a1a;
+      letter-spacing: 0.5px;
+    }
+  }
+
+  .currency-card-rates {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+    margin-left: auto;
+
+    .rate-value {
+      font-family: 'Inter', 'Segoe UI', sans-serif;
+      font-size: 1rem;
+      font-weight: 500;
+      color: #333;
+      line-height: 1.2;
+
+      &.buy {
+        color: #1a1a1a;
+      }
+
+      &.sell {
+        color: #666;
+      }
+    }
+  }
+}
+
+// Support for dark mode if root has a class
+.dark-theme .currency-card {
+  background: rgba(30, 30, 30, 0.6);
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.3);
+
+  .currency-code {
+    color: #fff;
+  }
+
+  .rate-value {
+    color: #ddd;
+
+    &.buy {
+      color: #fff;
+    }
+
+    &.sell {
+      color: #aaa;
+    }
+  }
+
+  .currency-flag-circle {
+    background: #333;
+  }
+}
