+++ src/styles/components/Menu.scss
@@ -1009,85 +1009,6 @@
   }
 }
 
-// Currency Rates Widget
-.currency-widget {
-  margin: 8px 0;
-  padding: 10px 12px;
-  background: #fafafa;
-  border-radius: 10px;
-  border: 1px solid #eee;
-  flex-shrink: 0;
-
-  .currency-widget-title {
-    font-size: 0.82rem;
-    font-weight: 600;
-    color: #555;
-    margin-bottom: 8px;
-    font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-  }
-
-  .currency-widget-loading,
-  .currency-widget-error {
-    font-size: 0.78rem;
-    color: #999;
-    text-align: center;
-    padding: 6px 0;
-  }
-
-  .currency-widget-error {
-    color: #e31919;
-  }
-
-  .currency-widget-table {
-    display: flex;
-    flex-direction: column;
-    gap: 4px;
-  }
-
-  .currency-widget-header {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    gap: 4px;
-    font-size: 0.68rem;
-    font-weight: 600;
-    color: #aaa;
-    text-transform: uppercase;
-    letter-spacing: 0.3px;
-    padding-bottom: 4px;
-    border-bottom: 1px solid #eee;
-  }
-
-  .currency-widget-row {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    gap: 4px;
-    padding: 4px 0;
-    transition: background 0.2s;
-    border-radius: 4px;
-
-    &:hover {
-      background: rgba(227, 25, 25, 0.04);
-    }
-  }
-
-  .currency-widget-name {
-    font-size: 0.82rem;
-    font-weight: 600;
-    color: #333;
-  }
-
-  .currency-widget-buy {
-    font-size: 0.82rem;
-    font-weight: 700;
-    color: #22c55e;
-  }
-
-  .currency-widget-sell {
-    font-size: 0.82rem;
-    font-weight: 700;
-    color: #e31919;
-  }
-}
 
 .sidebar-toggle {
   display: flex;
